feat: One more revised idea, this time low risk in edx platform and also helps address enterprise specific flow. testing pending
