+++ openedx/core/djangoapps/enrollments/views.py
@@ -6,6 +6,8 @@ consist primarily of authentication, request validation, and serialization.
 
 
 import logging
+from openedx.core.djangoapps.enrollments.exceptions import
+from openedx.core.djangoapps.enrollments.utils import add_user_to_course_cohort, check_mode_and_enroll
 
 from common.djangoapps.course_modes.models import CourseMode
 from django.core.exceptions import ObjectDoesNotExist, ValidationError  # lint-amnesty, pylint: disable=wrong-import-order
@@ -20,7 +22,7 @@ from openedx.core.djangoapps.course_groups.cohorts import CourseUserGroup, add_u
 from openedx.core.djangoapps.embargo import api as embargo_api
 from openedx.core.djangoapps.enrollments import api
 from openedx.core.djangoapps.enrollments.errors import (
-    CourseEnrollmentError, CourseEnrollmentExistsError, CourseModeNotFoundError,
+    CourseEnrollmentError, CourseEnrollmentExistsError, CourseModeNotFoundError, EnrollmentModeMismatchError, EnrollmentAttributesMissingError,
 )
 from openedx.core.djangoapps.enrollments.forms import CourseEnrollmentsApiListForm
 from openedx.core.djangoapps.enrollments.paginators import CourseEnrollmentsApiListPagination
@@ -50,9 +52,6 @@ from common.djangoapps.student.roles import CourseStaffRole, GlobalStaff
 from common.djangoapps.util.disable_rate_limit import can_disable_rate_limit
 
 log = logging.getLogger(__name__)
-REQUIRED_ATTRIBUTES = {
-    "credit": ["credit:provider_id"],
-}
 
 
 class EnrollmentCrossDomainSessionAuth(SessionAuthenticationAllowInactiveUser, SessionAuthenticationCrossDomainCsrf):
@@ -738,8 +737,10 @@ class EnrollmentListView(APIView, ApiKeyPermissionMixIn):
                     }
                 )
 
+            # Enterprise-specific enrollment steps
             explicit_linked_enterprise = request.data.get('linked_enterprise_customer')
             if explicit_linked_enterprise and has_api_key_permissions and enterprise_enabled():
+                # TODO : Refactor this piece to invoke the now available function in enterprise_support/utils
                 enterprise_api_client = EnterpriseApiServiceClient()
                 consent_client = ConsentApiServiceClient()
                 try:
@@ -755,62 +756,16 @@ class EnrollmentListView(APIView, ApiKeyPermissionMixIn):
                 }
                 consent_client.provide_consent(**kwargs)
 
+            # Mode changes etc
             enrollment_attributes = request.data.get('enrollment_attributes')
             enrollment = api.get_enrollment(username, str(course_id))
-            mode_changed = enrollment and mode is not None and enrollment['mode'] != mode
-            active_changed = enrollment and is_active is not None and enrollment['is_active'] != is_active
-            missing_attrs = []
-            if enrollment_attributes:
-                actual_attrs = [
-                    "{namespace}:{name}".format(**attr)
-                    for attr in enrollment_attributes
-                ]
-                missing_attrs = set(REQUIRED_ATTRIBUTES.get(mode, [])) - set(actual_attrs)
-            if (GlobalStaff().has_user(request.user) or has_api_key_permissions) and (mode_changed or active_changed):
-                if mode_changed and active_changed and not is_active:
-                    # if the requester wanted to deactivate but specified the wrong mode, fail
-                    # the request (on the assumption that the requester had outdated information
-                    # about the currently active enrollment).
-                    msg = "Enrollment mode mismatch: active mode={}, requested mode={}. Won't deactivate.".format(
-                        enrollment["mode"], mode
-                    )
-                    log.warning(msg)
-                    return Response(status=status.HTTP_400_BAD_REQUEST, data={"message": msg})
+            if (GlobalStaff().has_user(request.user) or has_api_key_permissions):
+                response = check_mode_and_enroll(username, course_id, mode, enrollment, enrollment_attributes, is_active, has_api_key_permissions)
 
-                if missing_attrs:
-                    msg = "Missing enrollment attributes: requested mode={} required attributes={}".format(
-                        mode, REQUIRED_ATTRIBUTES.get(mode)
-                    )
-                    log.warning(msg)
-                    return Response(status=status.HTTP_400_BAD_REQUEST, data={"message": msg})
-
-                response = api.update_enrollment(
-                    username,
-                    str(course_id),
-                    mode=mode,
-                    is_active=is_active,
-                    enrollment_attributes=enrollment_attributes,
-                    # If we are updating enrollment by authorized api caller, we should allow expired modes
-                    include_expired=has_api_key_permissions
-                )
-            else:
-                # Will reactivate inactive enrollments.
-                response = api.add_enrollment(
-                    username,
-                    str(course_id),
-                    mode=mode,
-                    is_active=is_active,
-                    enrollment_attributes=enrollment_attributes
-                )
-
+            # Add to cohort as necessary
             cohort_name = request.data.get('cohort')
-            if cohort_name is not None:
-                cohort = get_cohort_by_name(course_id, cohort_name)
-                try:
-                    add_user_to_cohort(cohort, user)
-                except ValueError:
-                    # user already in cohort, probably because they were un-enrolled and re-enrolled
-                    log.exception('Cohort re-addition')
+            add_user_to_course_cohort(cohort_name, course_id, user)
+
             email_opt_in = request.data.get('email_opt_in', None)
             if email_opt_in is not None:
                 org = course_id.org
@@ -856,6 +811,8 @@ class EnrollmentListView(APIView, ApiKeyPermissionMixIn):
                 data={
                     "message": "An error occured while adding to cohort [%s]" % cohort_name
                 })
+        except (EnrollmentModeMismatchError, EnrollmentAttributesMissingError) as error:
+            return Response(status=status.HTTP_400_BAD_REQUEST, data={"message": str(error)})
         finally:
             # Assumes that the ecommerce service uses an API key to authenticate.
             if has_api_key_permissions:
