+++ openedx/features/enterprise_support/enrollments/exceptions.py
@@ -10,15 +10,3 @@ class UserDoesNotExistException(Exception):
     """
     course_id invalid
     """
-
-
-class EnrollmentModeMismatchError(Exception):
-    """
-    requester has outdated information about the currently active enrollment
-    """
-
-
-class EnrollmentAttributesMissingError(Exception):
-    """
-    some enrollment attributes are missing
-    """
